feat(dilesa-drawers-a-paginas): Sprint 1 — drawers DILESA → páginas completas
Iniciativa nueva `dilesa-drawers-a-paginas`. Promovida + ejecutada +
cerrada en 1 PR (Sprint 1 + closeout).

Beto pidió: "quitar todos los side drawers de DILESA con excepción
del de Tareas". Verificación inicial: DILESA tenía solo 2 detail
drawers efectivos (Proyecto y Anteproyecto).

Refactor in-place:

- `components/dilesa/proyecto-detail-drawer.tsx` → `proyecto-detalle.tsx`
  (git mv). El export `<ProyectoDetailDrawer>` se convierte en
  `<ProyectoDetalle>` (signature simplificada: solo recibe `proyecto`,
  sin `open`/`onOpenChange`). Quita el wrapping `<DetailDrawer>`/
  `<DetailDrawerContent>` por `<div className="space-y-6 p-4 sm:p-6">`
  con `<header>` para título + badges de tipo y estado. Conserva
  todo el resto (types, labels, effects, edición inline, sección
  Avances, sección Documentos y configuración del Sprint A de
  `dilesa-proyectos-paridad-coda`).
- Mismo refactor para `anteproyecto-detail-drawer.tsx` →
  `anteproyecto-detalle.tsx` con `<AnteproyectoDetalle>`. Conserva
  sección de Checklist, Presupuestos preliminares y Promoción a
  desarrollo del Sprint 4 de la iniciativa hermana.

Pages nuevas:

- `app/dilesa/proyectos/[id]/page.tsx`: hereda sub-slug
  `dilesa.proyectos.activos`. Fetch del proyecto + breadcrumb
  "Volver a proyectos" + monta `<ProyectoDetalle>`. Estado `loading`
  y `notFound` apropiado.
- `app/dilesa/proyectos/anteproyectos/[id]/page.tsx`: hereda sub-slug
  `dilesa.proyectos.anteproyectos`. Filtra `tipo='anteproyecto'` en el
  query + breadcrumb "Volver a anteproyectos" + monta
  `<AnteproyectoDetalle>`.

Listas: `<ProyectosModule>` y `<AnteproyectosModule>` cambian
`onRowClick` de `setSelected + setDrawerOpen` a
`router.push('/dilesa/proyectos/[id]')` /
`router.push('/dilesa/proyectos/anteproyectos/[id]')`. Eliminado
state `selected`/`drawerOpen`. Deep-linking automático.

Decisiones registradas (DP1-DP6):

- DP1: el detalle es una page, no un drawer.
- DP2: breadcrumb "Volver" con `<Link>` canónico al listado padre.
- DP3: secciones verticales scroll-largo (no sub-tabs hasta ~5
  secciones grandes).
- DP4: edición inline en cada sección.
- DP5: `<RequireAccess>` + `<DesktopOnlyNotice>` por defecto.
- DP6: `<TasksUpdates>` única excepción permitida como drawer
  (contextual a la entidad principal).

Sprint 2 NO REQUERIDO: verificación al cierre confirma que el resto
de módulos DILESA (Ventas, Construcción, Inventario, Estimaciones,
Contratos, Contratistas, Prototipos, Portafolio) ya operan en pages
directamente o en módulos con tabs. Ningún módulo DILESA usa
`<DetailDrawer>` además de los 2 ya convertidos.

924 tests verdes. Cero cambios en factories de tests — solo refactor
de componentes existentes preservando lógica.
